fix the benchmarks to be more accurate
previously returning the object was probably making copies; returning
the pointer seems to suffice as the compiler (or at least msvc) doesn't
seem to optimise away the array calculations. this is evidenced by the
timer increasing 10x when the length is 10x.

in this new comparison, eigen is equivalent to IPP.